User registration and courses (#7)
* midi_device for synth

Uploaded updated midi_device.js to be used with internal synth

* dependencies for synth

Tone.js and Teoria.js (used to facilitate synth)

* Add files via upload

Added a courses folder with a folder with course_id 1.
This can be accessed through the manage view or the new courses view

* Added admin and register to project urls.py

Added admin and register to project urls.py

* Add files via upload

Changes in existing files to add navbar and welcome text. Addition of templates for registration etc. Changes in __base.html and additions of bases for course list and registration pages.

* Changes and additions to templates

Changes in existing files to add navbar and welcome text. Addition of templates for registration etc. Changes in __base.html and additions of bases for course list and registration pages.

* Lab urls.py and views.py

Added a course list view - this view is similar to manage view. It doesn't require special permissions to view and does not have the link to 'create exercise' - which doesn't work anyway.

* Added course_list view

This probably should move into lab views and urls but for some reason I added it here. It seems to work for now.

* Create requirements.txt

* Register app

An app that handles registration of new users

* Templates for registration 

Templates for registration

* Create __init__.py

* uploaded pychace of migrations

* pycache for register folder

* Update __base.html

* Update base.html

* Update base.html

* Update base_generic.html

* Update base.html

Co-authored-by: ospreyelm <rpmoseley@gmail.com>