Added alert to initialize web audio
If the audio context has not started, this will trigger an alert asking the user to click ok which will start the audio context. This only happens once per load and does not reappear on refresh. Need to test if this is true even inside of exercises on the server (I believe it will be)